probably fixed sessions errors on api
